requirements.txt
github3.py==0.9.6
pygit2==0.26
colorama==0.3.9
